requirements.txt
requests==2.31.0
boto3==1.34.0
twilio==8.10.0
python-dotenv==1.0.0